% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devindocker-package.R
\docType{package}
\name{devindocker-package}
\alias{devindocker}
\alias{devindocker-package}
\title{devindocker: Help Develop a R Project Inside a Docker Container}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Help develop a R project inside a Docker container.
}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
