% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-stop-docker.R
\name{launch_proj_docker}
\alias{launch_proj_docker}
\alias{stop_proj_docker}
\title{Launch and stop Docker}
\usage{
launch_proj_docker(
  project_path = "",
  container = "thinkr/rstudio3_6_1_geo",
  with_mysql = FALSE,
  mysql_docker = "mysql:8.0.16",
  with_chrome = FALSE,
  chrome_docker = "tkp1n/chromium",
  chrome_path = pagedown::find_chrome(),
  port = 8787,
  renv_inst = FALSE,
  renv_cache,
  update_docker = TRUE
)

stop_proj_docker(
  project_path,
  with_mysql = FALSE,
  with_chrome = FALSE,
  sleep = 10
)
}
\arguments{
\item{project_path}{Path to project to launch.}

\item{container}{Docker container to download from docker hub.}

\item{with_mysql}{Logical. Use MySQL database or not.}

\item{mysql_docker}{Mysql docker container to download from docker hub.}

\item{with_chrome}{Logical. Use Chrome in the network or not.}

\item{chrome_docker}{Chrome docker container to download from docker hub.}

\item{chrome_path}{Path to local chromium installation.
Help with \code{\link[pagedown:find_chrome]{pagedown::find_chrome()}}, but use the real path, not a symlink}

\item{port}{Local port to which to launch Rstudio Server}

\item{renv_inst}{Logical. Whether to add a R script with {renv} instructions in the project.}

\item{renv_cache}{Path to renv cache on your computer. Set to FALSE to not use renv.}

\item{update_docker}{Logical. Whether to update Docker container with DockerHub.}

\item{sleep}{Numeric. Number of seconds to wait for user to correctly stop Rstudio Server}
}
\description{
Launch and stop Docker

Stop running Docker container
}
\examples{
\dontrun{
tempdir <- tempdir()
project_path <- file.path(tempdir, "myproject")
usethis::create_package(project_path, open = FALSE)

# Launch Docker

# Which Rstudio container ? ----
container <- c("thinkr/rstudio3_5_2",
               "thinkr/rstudio3_5_2_geo",
               "thinkr/rstudio3_6_1_geo")[3]

# Which port ? ----
# _Useful if multiple Rstudio Server to launch
port <- 8788

# Start Docker project
launch_proj_docker(project_path = project_path,
                   container = container,
                   port = port)

# Stop Docker properly
stop_proj_docker(project_path = project_path, sleep = 5)

# With renv cache shared with host
dir.create(file.path(tempdir, "cache"))
tempcache <- file.path(tempdir, "cache")
# Start Docker project
launch_proj_docker(project_path = project_path,
                   container = container,
                   port = port,
                   renv_cache = tempcache)

# Stop Docker properly
stop_proj_docker(project_path = project_path, sleep = 5)
}
}
