% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-stop-docker.R
\name{launch_proj_docker}
\alias{launch_proj_docker}
\alias{stop_proj_docker}
\title{Launch and stop Docker}
\usage{
launch_proj_docker(
  path = ".",
  container = "thinkr/rstudio3_6_1_geo",
  network_name = "r-db",
  port = 8787,
  renv_inst = FALSE,
  renv_cache = FALSE,
  renv_out = FALSE,
  renv_out_dir,
  update_docker = TRUE,
  is_root = FALSE
)

stop_proj_docker(path, sleep = 10, network_name = "r-db", stop_network = TRUE)
}
\arguments{
\item{path}{Path to project or any directory to launch.}

\item{container}{Docker container to download from docker hub.}

\item{network_name}{Character. Give the name of the network in which the
container will be included using \code{--net network_name}.}

\item{port}{Local port to which to launch Rstudio Server}

\item{renv_inst}{Logical. Whether to add a R script with {renv} instructions in the project.}

\item{renv_cache}{Path to renv cache on your computer. Set to FALSE to not use renv.}

\item{renv_out}{Whether to set {renv} libraries out of the project.}

\item{renv_out_dir}{Where to store project libraries. Default to a ".renv" folder
in the parent directory of "path"}

\item{update_docker}{Logical. Whether to update Docker container with DockerHub.}

\item{is_root}{logical. Whether the Docker user has root permissions (add to sudoers).
Can be useful if you want to simulate your CI behaviour in the Terminal using "\code{sudo R}".}

\item{sleep}{Numeric. Number of seconds to wait for user to correctly stop Rstudio Server}

\item{stop_network}{Logical. Whether to stop Docker network.}
}
\description{
Launch and stop Docker

Stop running Docker container
}
\details{
renv_out : It is recommended to set {renv library} out of the project to avoid copy of entire library
during package checks. See \code{vignette("packages", package = "renv")}.
}
\examples{
\dontrun{
tempdir <- tempdir()
path <- file.path(tempdir, "myproject")
usethis::create_package(path, open = FALSE)

# Launch Docker

# Which Rstudio container ? ----
container <- c("thinkr/rstudio3_5_2",
               "rocker/geospatial:4.0.1")[2]

# Which port ? ----
# _Useful if multiple Rstudio Server to launch
port <- 8788

# Start Docker project
launch_proj_docker(path = path,
                   container = container,
                   port = port)

# Stop Docker properly
stop_proj_docker(path = path, sleep = 5)

# With renv cache shared with host
dir.create(file.path(tempdir, "cache"))
tempcache <- file.path(tempdir, "cache")
# Start Docker project
launch_proj_docker(path = path,
                   container = container,
                   port = port,
                   renv_cache = tempcache)

# Stop Docker properly
stop_proj_docker(path = path, sleep = 5)
}
}
